Add "clickToLoad" feature configuration and deprecate "clickToPlay" (#721)
The Click to Load feature was misnamed "clickToPlay" in the extension
configuration, which had a knock-on effect with the naming of files in
the content-scope-scripts and other repositories. Let's fix that now,
or at least start the towards fixing that.

We can't remove the old "clickToPlay" feature configuration yet, but
we can deprecate it and add the new "clickToLoad" feature
configuration ready for use. Also, let's take care to remove the old
"clicksBeforeSimpleVersion" setting from the new configuration and
correct the Facebook entity name ("Facebook, Inc." instead of
"Facebook") to match our block list.