Disable YouTube "Click to Load" again
While not as bad as last time, we still are receiving quite a few
breakage reports relating to the YouTube "Click to Load" feature. Some
seem to be breakage, but some seem to be that the user didn't like the
feature. In any case, let's disable the feature again now while we
investigate further.