Hide empty whitespace in footer of asuracomit.net
For some reason, the `.h12container` element that sometimes contains
adverts at the footer of the page is displayed as a white block that
users complain breaks the site. Hiding that seems to fix the issue,
note that it doesn't hide the advert contained, but it looks less
broken and seems to help fix the close advert button.

Note: I initially tried adding a "hide-empty" rule, but it didn't seem
      to be enough here.