Fix pageload issue on itsthevibe.com
It turns out that some of the pages on
itsthevibe.com (e.g. https://itsthevibe.com/most-loyal-dog-breeds/) fail to load
when these requests are blocked. It seems that a value that they set on the
"FCCDCF" property of some Object is later read back, if that fails the page
stops loading. Let's unblock those for now, to get them working again.

Note: There are also some rendering issues in the DuckDuckGo iOS browser, but
      those are also present in Safari.