Avoid using "internal" state for Windows browser where unsupported
The current version of the Windows browser uses the v3 configuration and does
not support the "internal" state for features or sub-features. If that state is
given, the browser fails to parse the configuration. In the near future, the
Windows browser will add support for "internal" and move to the v4
configuration. In the mean time, let's ensure that "internal" will be ready for
use in the Windows browser by replacing it with "disabled" where it isn't
supported.