Fix missing article comments on nationalpost.com (#3969)
Some users complained that the article comments on nationalpost.com were
missing. Looking into it, that seemed to be due to a couple of
blocked/redirected requests. Let's allow those requests for now so the comments
show up again.

While we're at it, let's fix a Tracker Allowlisting rule for a similar request
that incorrectly uses the .* wildcard syntax. So far, wildcards are not
supported here.