Add initial Request Blocklist feature configuration (#3990)
Add an initial configuration for the new Request Blocklist feature. Include a
couple of testing rules, for the privacy test page[1]. Also include a
corresponding Tracker Allowlist rule for the test page as well.

The feature is only implemented for the extension so far, so let's enable it
there but leave it disabled for other platforms.

1 - https://privacy-test-pages.site/privacy-protections/request-blocklist/