Workaround redirect-loop issue on cnbc.com (#4151)
There is an issue today on cnbc.com which causes the page to reload after the
"Reject All" button is clicked in the cookie prompt. When combined with Cookie
Prompt Management (CPM) that automatically clicks that button, users were seeing
the page reload over and over again. Let's work around the issue for now.