Hide empty space on msn.com (#4154)
When viewing articles, if the top banner advert doesn't load it leaves a large
empty space. That seems to be giving the impression to users that the page
hasn't finished loading, causing them to refresh repeatedly trying to load the
article. Let's collapse the element in question when it's empty.